requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.35.0  # For tokenizer compatibility
tokenizers>=0.15.0   # For training custom tokenizer

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0  # Optional: for experiment tracking
python-dotenv>=1.0.0

# Logging and Monitoring
tensorboard>=2.14.0
